put box label identifier and shipment query to global file
