+++ react/src/queries/queries.ts
@@ -1,5 +1,11 @@
 import { gql } from "@apollo/client";
-import { SHIPMENT_FIELDS_FRAGMENT, BOX_FIELDS_FRAGMENT } from "./fragments";
+import {
+  SHIPMENT_FIELDS_FRAGMENT,
+  BOX_FIELDS_FRAGMENT,
+  PRODUCT_BASIC_FIELDS_FRAGMENT,
+  TAG_FIELDS_FRAGMENT,
+  DISTRO_EVENT_FIELDS_FRAGMENT,
+} from "./fragments";
 
 export const BOX_DETAILS_BY_LABEL_IDENTIFIER_QUERY = gql`
   ${BOX_FIELDS_FRAGMENT}
@@ -31,6 +37,59 @@ export const CHECK_IF_QR_EXISTS_IN_DB = gql`
 export const ALL_SHIPMENTS_QUERY = gql`
   ${SHIPMENT_FIELDS_FRAGMENT}
   query Shipments {
+    shipments {
+      ...ShipmentFields
+    }
+  }
+`;
+
+export const BOX_BY_LABEL_IDENTIFIER_AND_ALL_SHIPMENTS_QUERY = gql`
+  ${PRODUCT_BASIC_FIELDS_FRAGMENT}
+  ${BOX_FIELDS_FRAGMENT}
+  ${TAG_FIELDS_FRAGMENT}
+  ${DISTRO_EVENT_FIELDS_FRAGMENT}
+  ${SHIPMENT_FIELDS_FRAGMENT}
+  query BoxByLabelIdentifier($labelIdentifier: String!) {
+    box(labelIdentifier: $labelIdentifier) {
+      ...BoxFields
+      product {
+        ...ProductBasicFields
+      }
+      tags {
+        ...TagFields
+      }
+      distributionEvent {
+        ...DistroEventFields
+      }
+      location {
+        __typename
+        id
+        name
+        ... on ClassicLocation {
+          defaultBoxState
+        }
+        base {
+          locations {
+            id
+            seq
+            name
+            ... on ClassicLocation {
+              defaultBoxState
+            }
+          }
+          distributionEventsBeforeReturnedFromDistributionState {
+            id
+            state
+            distributionSpot {
+              name
+            }
+            name
+            plannedStartDateTime
+            plannedEndDateTime
+          }
+        }
+      }
+    }
     shipments {
       ...ShipmentFields
     }
